+++ src/main/java/think/rpgitems/power/PowerLightning.java
@@ -47,7 +47,7 @@ public class PowerLightning extends Power implements PowerHit, PowerProjectileHi
 
     @Override
     public String displayText(String locale) {
-        return ChatColor.GREEN + String.format(Locale.get("power.lightning", locale), (int) ((1d / (double) chance) * 100d));
+        return ChatColor.GREEN + String.format(Locale.get("power.lightning", locale), (int) ((1d / chance) * 100d));
     }
 
     @Override
